Static IP assignment, private network interface configuration, and floating IP address allocation

+++ lib/vagrant-openstack-plugin/action/delete_server.rb
@@ -19,7 +19,16 @@ module VagrantPlugins
 
             # TODO: Validate the fact that we get a server back from the API.
             server = env[:openstack_compute].servers.get(id)
-            server.destroy
+            if server
+              ip = server.floating_ip_address
+              server.destroy
+              if machine.provider_config.floating_ip_pool
+                address = env[:openstack_compute].list_all_addresses.body["floating_ips"].find{|i| i["ip"] == ip}
+                if address
+                  env[:openstack_compute].release_address(address["id"])
+                end
+              end
+            end
           end
 
           @app.call(env)
